Parse JSON to struct...
Takes advantage of the rich facilities in crystal to map JSON to
known types. In order to be able to  parse, needs an initial known
key, so we provide one artifically.

{ "123": {}, "456": {} }

Becomes

{ "data": { "123": {}, "456": {} } }

This allows the first parsing to be of type Hash(String, Overdrive:Item).
Everything else is done by the mapping macro and type system