Remove generic type on ConcurrentSearch...
Instead of using a union between a generic type and the wanted type,
I defined a SearchInterface so that both the real Search and the
TestSearch can use it as a type. This necessitates adding the
as(SearchInterface) when instantiating a search implementation.

I think that is cleaner than having to add a generic, only so that
a double can be injected in the tests.